Added shader loading library and game directories.
The pull request adds the SimpleShaderLoading library made by me, along with a README.txt explaining it. It also adds directories in which game files can be placed, and also adds a lua script for the game. These updates were made to improve the management of both shaders and game files.